pyproject.toml
[project]
name = "twjobs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.0",
    "psycopg>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.46",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.14",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 79
extend-exclude = ["venv", ".venv", "__pycache__", "migrations"]

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "FAST"]

[tool.taskipy.tasks]
lint = "ruff check"
pre_format = "ruff check --fix"
format = "ruff format"
dev = "fastapi dev twjobs/app.py"
